Fixed a problem with world space triangle not updating dx1 and dx2 correctly after positions were changed. Still running into problems.

+++ VR Cloth Simulation - Project Pikachu/Packages/com.pika.simulation-helpers/Runtime/SimulationHelpers/Visualization/GraphVisualizer.cs
@@ -32,7 +32,7 @@ namespace SimulationHelpers.Visualization
             runStatistics = new List<IRunStatistic>();
         }
 
-        public override void Visualize(List<Vector3> positions, float elapsed, float dt)
+        public override void Visualize(List<Vector3> positions, double elapsed, double dt)
         {
             if (expectedPositionCount is null) 
                 expectedPositionCount = positions.Count;
